+++ CHANGELOG.md
@@ -2,14 +2,17 @@
 
 All Notable changes to `support` will be documented in this file
 
-## 1.2.0 - 2016-08-22
+## 1.2.0 - 2016-08-24
 
 - ADD IP helpers.
 - ADD getClientIp() by small refactor of Synfony functions with unit test.
-- ADD xml2array() and array2xml().
 - CHANGE mark OBSOLETE getIPVisitor().
+- ADD isVATNumber().Validate a European VAT number using the EU commission VIES service with soap.
+- ADD xml2array() and array2xml().
+- ADD generateRandomPassword().
 - ADD some new helpers.
 - ADD sensiolabs_security_checker in scrutinizer config.
+- FIX in xmlUrl2array.
 - COMMENT ocular code-coverage:upload to scritinizer in travis config.
 
 ## 1.1.0 - 2016-08-21